refactor(plateform): use shared api helper in mission-browse
